+++ model/src/main/java/org/xavier/common/exception/Universal_403_X_Exception_Runtime.java
@@ -15,27 +15,27 @@ import org.xavier.common.exception.base.RequestException_Runtime;
 public class Universal_403_X_Exception_Runtime extends RequestException_Runtime {
     public Universal_403_X_Exception_Runtime() {
         super();
-        setStateCode(403d);
+        setStateCode(403F);
     }
 
     public Universal_403_X_Exception_Runtime(String msg) {
         super(msg);
-        setStateCode(403d);
+        setStateCode(403F);
     }
 
-    public Universal_403_X_Exception_Runtime(Double stateCode, String msg) {
+    public Universal_403_X_Exception_Runtime(Float stateCode, String msg) {
         super(stateCode, msg);
     }
 
-    public Universal_403_X_Exception_Runtime(Double stateCode, String msg, Throwable cause) {
+    public Universal_403_X_Exception_Runtime(Float stateCode, String msg, Throwable cause) {
         super(stateCode, msg, cause);
     }
 
-    public Universal_403_X_Exception_Runtime(Double stateCode, String msg, String extraInfo) {
+    public Universal_403_X_Exception_Runtime(Float stateCode, String msg, String extraInfo) {
         super(stateCode, msg, extraInfo);
     }
 
-    public Universal_403_X_Exception_Runtime(Double stateCode, String msg, String extraInfo, Throwable cause) {
+    public Universal_403_X_Exception_Runtime(Float stateCode, String msg, String extraInfo, Throwable cause) {
         super(stateCode, msg, extraInfo, cause);
     }
 }
