+++ model/src/main/java/org/xavier/common/exception/Universal_404_X_Exception.java
@@ -14,26 +14,26 @@ import org.xavier.common.exception.base.RequestException;
 public class Universal_404_X_Exception extends RequestException {
     public Universal_404_X_Exception() {
         super();
-        setStateCode(404d);
+        setStateCode(404F);
     }
 
     public Universal_404_X_Exception(String msg) {
         super(msg);
-        setStateCode(404d);
+        setStateCode(404F);
     }
 
-    public Universal_404_X_Exception(Double stateCode, String msg) {
+    public Universal_404_X_Exception(Float stateCode, String msg) {
         super(stateCode, msg);
     }
 
-    public Universal_404_X_Exception(Double stateCode, String msg, Throwable cause) {
+    public Universal_404_X_Exception(Float stateCode, String msg, Throwable cause) {
         super(stateCode, msg, cause);
     }
-    public Universal_404_X_Exception(Double stateCode, String msg, String extraInfo) {
+    public Universal_404_X_Exception(Float stateCode, String msg, String extraInfo) {
         super(stateCode, msg, extraInfo);
     }
 
-    public Universal_404_X_Exception(Double stateCode, String msg, Throwable cause, String extraInfo) {
+    public Universal_404_X_Exception(Float stateCode, String msg, Throwable cause, String extraInfo) {
         super(stateCode, msg, extraInfo, cause);
     }
